requirements.txt
-i https://mirrors.aliyun.com/pypi/simple
# Flask
Flask == 1.0.3
Werkzeug == 0.15.4
click >= 5.0
# Api
Flask-RESTful == 0.3.7
# Database
Flask-SQLAlchemy == 2.4.0
SQLAlchemy == 1.3.5
PyMySQL == 0.9.3
redis == 3.2.1
# Migrations
Flask-Migrate == 2.5.2
# Deployment
gevent == 1.4.0
gunicorn == 19.5.0
supervisor == 4.0.3
# Auth
Flask-Login == 0.4.1
Flask-Bcrypt == 0.7.1
Flask-Cors >= 3.0.8
# Caching
Flask-Caching >= 1.0.0
# Environment variable parsing
environs == 4.2.0
marshmallow == 2.20.2
# async tasks
celery == 4.3.0
more-itertools == 5.0.0
kombu == 4.4.0
# other
six == 1.12.0
bs4 >= 0.0.1
toposort >= 1.5
requests >= 2.22.0
PyJWT >= 1.7.1
elasticsearch == 7.0.4
future == 0.18.2
paramiko==2.6.0
cffi==1.14.0